+++ modules/akai-fire/src/test/java/com/oikoaudio/fire/perform/PerformTrackControlOverlayTest.java
@@ -59,4 +59,5 @@ class PerformTrackControlOverlayTest {
         assertEquals(BiColorLightState.OFF,
                 PerformClipLauncherMode.patternSceneNavigationLightState(true, true));
     }
+
 }